Delete and edit playlists with old view

+++ app/src/main/java/ch/zhaw/engineering/tbdappname/services/database/dto/PlaylistWithSongCount.java
@@ -1,11 +1,10 @@
 package ch.zhaw.engineering.tbdappname.services.database.dto;
 
-import ch.zhaw.engineering.tbdappname.services.database.entity.Playlist;
 import lombok.Data;
 
 @Data
 public class PlaylistWithSongCount {
-    long playlistId;
+    int playlistId;
     String name;
     int songCount;
 }
